Move ccf out of tests
